# artifacts/api-server/src/routes/routers.ts
import { Router } from "express";
import { eq } from "drizzle-orm";
import { db, routersTable } from "@workspace/db";
import {
  ListRoutersQueryParams,
  ListRoutersResponse,
  CreateRouterBody,
  GetRouterParams,
  GetRouterResponse,
  UpdateRouterParams,
  UpdateRouterBody,
  UpdateRouterResponse,
  DeleteRouterParams,
  GetRouterRscParams,
  GetRouterRscResponse,
  GetRouterStatusParams,
  GetRouterStatusResponse,
} from "@workspace/api-zod";

const router = Router();

function generateRscScript(r: { name: string; ipAddress: string; apiPort: number; username: string; orgId: number }): string {
  const apiUrl = `${process.env.REPLIT_DOMAINS?.split(",")[0] ?? "fastanet.example.com"}/api`;
  return `# FASTANET Configuration Script for ${r.name}
# Generated by FASTANET ISP Billing System
# Router: ${r.name} | IP: ${r.ipAddress}

:global version [/system package update get installed-version];
:local majorVersion 0;
:local minorVersion 0;
:local dotPos [:find $version "."];
:if ([:len $dotPos] > 0) do={
    :set majorVersion [:tonum [:pick $version 0 $dotPos]];
    :local remaining [:pick $version ($dotPos + 1) [:len $version]];
    :set dotPos [:find $remaining "."];
    :if ([:len $dotPos] > 0) do={
        :set minorVersion [:tonum [:pick $remaining 0 $dotPos]];
    }
}
:if ($majorVersion < 6 || ($majorVersion = 6 && $minorVersion < 49)) do={
    :put "RouterOS version 6.49 or higher is required.";
    :error "RouterOS version 6.49 or higher is required.";
}
:if ([/ping 8.8.8.8 count=3] = 0) do={
    :error "No internet connection. Please check your internet connection and try again.";
}

# FASTANET API Configuration
:local fastanetApiUrl "${apiUrl}";
:local fastanetOrgId "${r.orgId}";
:local fastanetApiPort "${r.apiPort}";

# Configure API access
/ip service
:if ([find name=api] = "") do={
    add name=api port=${r.apiPort} disabled=no;
} else={
    set [find name=api] port=${r.apiPort} disabled=no;
}

# Configure hotspot profile for FASTANET
/ip hotspot profile
:if ([find name=fastanet] = "") do={
    add name=fastanet hotspot-address=${r.ipAddress} login-by=http-chap,http-pap smtp-server=0.0.0.0 html-directory=hotspot;
}

# Configure PPPoE server profile
/ppp profile
:if ([find name=fastanet-pppoe] = "") do={
    add name=fastanet-pppoe use-ipv6=no use-mpls=no use-compression=no use-encryption=no;
}

# Enable API user
/user
:if ([find name="${r.username}"] = "") do={
    add name="${r.username}" group=full;
} else={
    set [find name="${r.username}"] group=full;
}

:put "FASTANET configuration applied successfully.";
:put "Router: ${r.name}";
:put "API Endpoint: $fastanetApiUrl";
:put "Please ensure API port ${r.apiPort} is open in /ip firewall filter";
`;
}

router.get("/routers", async (req, res): Promise<void> => {
  const queryParams = ListRoutersQueryParams.safeParse(req.query);
  let query = db.select().from(routersTable).$dynamic();
  if (queryParams.success && queryParams.data.orgId) {
    query = query.where(eq(routersTable.orgId, queryParams.data.orgId));
  }
  const routers = await query.orderBy(routersTable.createdAt);
  const result = routers.map(r => {
    const { passwordHash: _, ...rest } = r;
    return rest;
  });
  res.json(ListRoutersResponse.parse(result));
});

router.post("/routers", async (req, res): Promise<void> => {
  const parsed = CreateRouterBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: parsed.error.message });
    return;
  }
  const { password, ...rest } = parsed.data;
  const [r] = await db.insert(routersTable).values({ ...rest, passwordHash: password, apiPort: rest.apiPort ?? 8728, type: rest.type ?? "hotspot" }).returning();
  const { passwordHash: _, ...safeRouter } = r;
  res.status(201).json(GetRouterResponse.parse(safeRouter));
});

router.get("/routers/:id", async (req, res): Promise<void> => {
  const params = GetRouterParams.safeParse(req.params);
  if (!params.success) {
    res.status(400).json({ error: params.error.message });
    return;
  }
  const [r] = await db.select().from(routersTable).where(eq(routersTable.id, params.data.id));
  if (!r) {
    res.status(404).json({ error: "Router not found" });
    return;
  }
  const { passwordHash: _, ...safeRouter } = r;
  res.json(GetRouterResponse.parse(safeRouter));
});

router.get("/routers/:id/rsc", async (req, res): Promise<void> => {
  const params = GetRouterRscParams.safeParse(req.params);
  if (!params.success) {
    res.status(400).json({ error: params.error.message });
    return;
  }
  const [r] = await db.select().from(routersTable).where(eq(routersTable.id, params.data.id));
  if (!r) {
    res.status(404).json({ error: "Router not found" });
    return;
  }
  const script = generateRscScript(r);
  const filename = `fastanet-${r.name.toLowerCase().replace(/\s+/g, "-")}.rsc`;
  res.json(GetRouterRscResponse.parse({ routerId: r.id, routerName: r.name, script, filename }));
});

router.get("/routers/:id/status", async (req, res): Promise<void> => {
  const params = GetRouterStatusParams.safeParse(req.params);
  if (!params.success) {
    res.status(400).json({ error: params.error.message });
    return;
  }
  const [r] = await db.select().from(routersTable).where(eq(routersTable.id, params.data.id));
  if (!r) {
    res.status(404).json({ error: "Router not found" });
    return;
  }
  res.json(GetRouterStatusResponse.parse({
    routerId: r.id,
    status: r.status,
    uptimeSeconds: r.uptimeSeconds,
    activeUsers: r.activeUsers,
    cpuLoad: null,
    freeMemory: null,
    lastCheckedAt: new Date().toISOString(),
  }));
});

router.patch("/routers/:id", async (req, res): Promise<void> => {
  const params = UpdateRouterParams.safeParse(req.params);
  if (!params.success) {
    res.status(400).json({ error: params.error.message });
    return;
  }
  const parsed = UpdateRouterBody.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ error: parsed.error.message });
    return;
  }
  const { password, ...rest } = parsed.data as Record<string, unknown> & { password?: string };
  const updateData: Record<string, unknown> = { ...rest };
  if (password) updateData.passwordHash = password;
  const [r] = await db.update(routersTable).set(updateData).where(eq(routersTable.id, params.data.id)).returning();
  if (!r) {
    res.status(404).json({ error: "Router not found" });
    return;
  }
  const { passwordHash: _, ...safeRouter } = r;
  res.json(UpdateRouterResponse.parse(safeRouter));
});

router.delete("/routers/:id", async (req, res): Promise<void> => {
  const params = DeleteRouterParams.safeParse(req.params);
  if (!params.success) {
    res.status(400).json({ error: params.error.message });
    return;
  }
  const [r] = await db.delete(routersTable).where(eq(routersTable.id, params.data.id)).returning();
  if (!r) {
    res.status(404).json({ error: "Router not found" });
    return;
  }
  res.sendStatus(204);
});

export default router;