# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FASTANET ISP Billing & Hotspot Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /routers:
    get:
      operationId: listRouters
      tags: [routers]
      summary: List all MikroTik routers
      parameters:
        - name: orgId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of routers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Router"
    post:
      operationId: createRouter
      tags: [routers]
      summary: Add a MikroTik router
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRouterBody"
      responses:
        "201":
          description: Router added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Router"
  /routers/{id}:
    get:
      operationId: getRouter
      tags: [routers]
      summary: Get router by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Router
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Router"
    patch:
      operationId: updateRouter
      tags: [routers]
      summary: Update a router
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRouterBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Router"
    delete:
      operationId: deleteRouter
      tags: [routers]
      summary: Delete a router
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /routers/{id}/rsc:
    get:
      operationId: getRouterRsc
      tags: [routers]
      summary: Download MikroTik .rsc configuration script for this router
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: RSC script content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouterRscResponse"
  /routers/{id}/status:
    get:
      operationId: getRouterStatus
      tags: [routers]
      summary: Get router connection status and stats
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Router status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouterStatus"

  # ISP Users
...
  /vouchers:
    get:
      operationId: listVouchers
      tags: [vouchers]
      summary: List vouchers
      parameters:
        - name: orgId
          in: query
          required: false
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [unused, used, expired]
      responses:
        "200":
          description: List of vouchers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Voucher"
    post:
      operationId: generateVouchers
      tags: [vouchers]
      summary: Generate bulk vouchers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateVouchersBody"
      responses:
        "201":
          description: Vouchers generated
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Voucher"
  /vouchers/{id}:
    get:
      operationId: getVoucher
      tags: [vouchers]
      summary: Get voucher by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Voucher
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Voucher"
    delete:
      operationId: deleteVoucher
      tags: [vouchers]
      summary: Delete a voucher
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # PPPoE Accounts
...
components:
  schemas:
...
    Router:
      type: object
      properties:
        id:
          type: integer
        orgId:
          type: integer
        name:
          type: string
        ipAddress:
          type: string
        apiPort:
          type: integer
        username:
          type: string
        model:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        status:
          type: string
          enum: [online, offline, unknown]
        uptimeSeconds:
          type: ["integer", "null"]
        activeUsers:
          type: ["integer", "null"]
        type:
          type: string
          enum: [hotspot, pppoe, both]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - orgId
        - name
        - ipAddress
        - apiPort
        - username
        - status
        - type
        - createdAt
    CreateRouterBody:
      type: object
      properties:
        orgId:
          type: integer
        name:
          type: string
        ipAddress:
          type: string
        apiPort:
          type: integer
        username:
          type: string
        password:
          type: string
        model:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        type:
          type: string
          enum: [hotspot, pppoe, both]
      required:
        - orgId
        - name
        - ipAddress
        - username
        - password
    UpdateRouterBody:
      type: object
      properties:
        name:
          type: string
        ipAddress:
          type: string
        apiPort:
          type: integer
        username:
          type: string
        password:
          type: string
        model:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        type:
          type: string
          enum: [hotspot, pppoe, both]
        status:
          type: string
          enum: [online, offline, unknown]
    RouterRscResponse:
      type: object
      properties:
        routerId:
          type: integer
        routerName:
          type: string
        script:
          type: string
        filename:
          type: string
      required:
        - routerId
        - routerName
        - script
        - filename
    RouterStatus:
      type: object
      properties:
        routerId:
          type: integer
        status:
          type: string
          enum: [online, offline, unknown]
        uptimeSeconds:
          type: ["integer", "null"]
        activeUsers:
          type: ["integer", "null"]
        cpuLoad:
          type: ["number", "null"]
        freeMemory:
          type: ["string", "null"]
        lastCheckedAt:
          type: string
          format: date-time
      required:
        - routerId
        - status
        - lastCheckedAt
...
    Voucher:
      type: object
      properties:
        id:
          type: integer
        orgId:
          type: integer
        packageId:
          type: integer
        code:
          type: string
        status:
          type: string
          enum: [unused, used, expired]
        usedAt:
          type: ["string", "null"]
          format: date-time
        usedByUserId:
          type: ["integer", "null"]
        expiryDate:
          type: ["string", "null"]
          format: date-time
        packageName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - orgId
        - packageId
        - code
        - status
        - createdAt
    GenerateVouchersBody:
      type: object
      properties:
        orgId:
          type: integer
        packageId:
          type: integer
        count:
          type: integer
        expiryDays:
          type: ["integer", "null"]
      required:
        - orgId
        - packageId
        - count
...