# lib/api-zod/src/generated/types/createRouterBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FASTANET ISP Billing & Hotspot Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateRouterBodyType } from "./createRouterBodyType";

export interface CreateRouterBody {
  orgId: number;
  name: string;
  ipAddress: string;
  apiPort?: number;
  username: string;
  password: string;
  /** @nullable */
  model?: string | null;
  /** @nullable */
  location?: string | null;
  type?: CreateRouterBodyType;
}