# lib/api-zod/src/generated/types/router.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FASTANET ISP Billing & Hotspot Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { RouterStatusProperty } from "./routerStatusProperty";
import type { RouterType } from "./routerType";

export interface Router {
  id: number;
  orgId: number;
  name: string;
  ipAddress: string;
  apiPort: number;
  username: string;
  /** @nullable */
  model?: string | null;
  /** @nullable */
  location?: string | null;
  status: RouterStatusProperty;
  /** @nullable */
  uptimeSeconds?: number | null;
  /** @nullable */
  activeUsers?: number | null;
  type: RouterType;
  createdAt: Date;
}