# lib/api-zod/src/generated/types/routerRscResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FASTANET ISP Billing & Hotspot Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface RouterRscResponse {
  routerId: number;
  routerName: string;
  script: string;
  filename: string;
}