# lib/api-zod/src/generated/types/routerStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FASTANET ISP Billing & Hotspot Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { RouterStatusStatus } from "./routerStatusStatus";

export interface RouterStatus {
  routerId: number;
  status: RouterStatusStatus;
  /** @nullable */
  uptimeSeconds?: number | null;
  /** @nullable */
  activeUsers?: number | null;
  /** @nullable */
  cpuLoad?: number | null;
  /** @nullable */
  freeMemory?: string | null;
  lastCheckedAt: Date;
}